Create the canonical paths for atmosphere
Oddly, clank wants these paths available before they are created by
roles. This appears to have something to do with `--workspace`,
however the vagrant BUILD_ENV has `JENKINS: false` and does not set
the JENKINS_WORKSPACE variable. It is likely because the line calling
out to `ratchet.py` is not inspected $? (or ratchet.py is *not*
bubbling up the exit code).